Cleaning
git-svn-id: file:///home/svn/shoola-omero/trunk@6040 6f2cb1de-eb0d-0410-b157-e593188b5901

+++ SRC/org/openmicroscopy/shoola/agents/imviewer/rnd/GraphicsPane.java
@@ -231,8 +231,8 @@ class GraphicsPane
     	 int knobHeight = domainSlider.getKnobHeight();
     	 int width = codomainSlider.getPreferredSize().width;
     	 double size[][] =
-         {{width, knobWidth/2, TableLayout.FILL, knobWidth/2},  // Columns
-          {knobHeight/2, TableLayout.FILL, (double) knobHeight/2, 
+         {{width, (double) knobWidth/2, TableLayout.FILL, (double) knobWidth/2},  // Columns
+          {(double) knobHeight/2, TableLayout.FILL, (double) knobHeight/2, 
         	 (double) knobHeight+2, TableLayout.PREFERRED, 5}}; // Rows
     	 p.setLayout(new TableLayout(size));
     	 p.add(codomainSlider, "0, 0, 0, 2");
